Move models and exceptions to common
