+++ common/build.gradle
@@ -1,0 +1,6 @@
+apply plugin: "org.springframework.boot"
+apply plugin: 'io.spring.dependency-management'
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+}
