+++ common/src/main/java/nl/tudelft/wdm/group1/common/CreditChangeInvalidException.java
@@ -1,0 +1,11 @@
+package nl.tudelft.wdm.group1.common;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.UNPROCESSABLE_ENTITY)
+public class CreditChangeInvalidException extends Exception {
+    public CreditChangeInvalidException(String message) {
+        super(message);
+    }
+}
