+++ common/src/main/java/nl/tudelft/wdm/group1/common/InsufficientCreditException.java
@@ -1,4 +1,4 @@
-package nl.tudelft.wdm.group1.users;
+package nl.tudelft.wdm.group1.common;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
