+++ common/src/main/java/nl/tudelft/wdm/group1/common/InvalidStockChangeException.java
@@ -1,11 +1,11 @@
-package nl.tudelft.wdm.group1.users;
+package nl.tudelft.wdm.group1.common;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.UNPROCESSABLE_ENTITY)
-public class CreditChangeInvalidException extends Exception {
-    public CreditChangeInvalidException(String message) {
+public class InvalidStockChangeException extends Exception {
+    public InvalidStockChangeException(final String message) {
         super(message);
     }
 }
