+++ common/src/main/java/nl/tudelft/wdm/group1/common/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package nl.tudelft.wdm.group1.payments;
+package nl.tudelft.wdm.group1.common;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
