+++ common/src/main/java/nl/tudelft/wdm/group1/common/User.java
@@ -1,4 +1,4 @@
-package nl.tudelft.wdm.group1.users;
+package nl.tudelft.wdm.group1.common;
 
 import java.util.UUID;
