+++ orders/src/main/java/nl/tudelft/wdm/group1/orders/events/Consumer.java
@@ -1,9 +1,9 @@
 package nl.tudelft.wdm.group1.orders.events;
 
+import nl.tudelft.wdm.group1.common.Order;
 import nl.tudelft.wdm.group1.common.OrdersTopics;
-import nl.tudelft.wdm.group1.orders.Order;
+import nl.tudelft.wdm.group1.common.ResourceNotFoundException;
 import nl.tudelft.wdm.group1.orders.OrderRepository;
-import nl.tudelft.wdm.group1.orders.ResourceNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.kafka.annotation.KafkaListener;
