+++ orders/src/main/java/nl/tudelft/wdm/group1/orders/events/Producer.java
@@ -1,7 +1,7 @@
 package nl.tudelft.wdm.group1.orders.events;
 
+import nl.tudelft.wdm.group1.common.Order;
 import nl.tudelft.wdm.group1.common.OrdersTopics;
-import nl.tudelft.wdm.group1.orders.Order;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
