+++ orders/src/test/java/nl/tudelft/wdm/group1/orders/OrdersApplicationTest.java
@@ -1,6 +1,8 @@
 package nl.tudelft.wdm.group1.orders;
 
 import com.jayway.jsonpath.JsonPath;
+import nl.tudelft.wdm.group1.common.Order;
+import nl.tudelft.wdm.group1.common.ResourceNotFoundException;
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Test;
