+++ payments/src/main/java/nl/tudelft/wdm/group1/payments/events/Consumer.java
@@ -1,8 +1,8 @@
 package nl.tudelft.wdm.group1.payments.events;
 
-import nl.tudelft.wdm.group1.payments.Payment;
+import nl.tudelft.wdm.group1.common.Payment;
+import nl.tudelft.wdm.group1.common.ResourceNotFoundException;
 import nl.tudelft.wdm.group1.payments.PaymentRepository;
-import nl.tudelft.wdm.group1.payments.ResourceNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.kafka.annotation.KafkaListener;
