+++ payments/src/main/java/nl/tudelft/wdm/group1/payments/web/PaymentController.java
@@ -1,16 +1,11 @@
 package nl.tudelft.wdm.group1.payments.web;
 
+import nl.tudelft.wdm.group1.common.Payment;
+import nl.tudelft.wdm.group1.common.ResourceNotFoundException;
 import nl.tudelft.wdm.group1.payments.PaymentRepository;
-import nl.tudelft.wdm.group1.payments.ResourceNotFoundException;
-import nl.tudelft.wdm.group1.payments.Payment;
 import nl.tudelft.wdm.group1.payments.events.Producer;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.UUID;
