--- stock/src/main/java/nl/tudelft/wdm/group1/stock/InsufficientStockException.java
@@ -1,9 +1,0 @@
-package nl.tudelft.wdm.group1.stock;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class InsufficientStockException extends Exception{
-    public InsufficientStockException(String message) { super(message); }
-}
