--- users/src/main/java/nl/tudelft/wdm/group1/users/ResourceNotFoundException.java
@@ -1,11 +1,0 @@
-package nl.tudelft.wdm.group1.users;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends Exception {
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
